stuck grabs
slime projectiles now can create a self grab on the player. enemys can grab the player into a new grab animation while they are self grabbed. allowing for the projectiles to be a decent obsticle. functality is there however, animations and gooed player need to be added for male/ created for female. then next is updating the blue slime enemy's sprite so that it is transparent.