fixed a painful bug
fixed a painful bug where a settings reload would cause the justdown functionality of our mobile buttons to stop working. it was a painful debugging process, but i opted to simply not restart the game hud and instead, just update it with new values after closing it when the reload goes through after pressing the yes button.