additions
added two new weapons and a ring, which can be obtained by defeating the mimics in the cave with the two slimes near the safehouse. and the ring can be obtained by defeating the mimic in the cave near the sunflower field. rapier is deals cursed damage but the ring does nothing currently